Refactoring and cleanup
- Moved the oglplus math utilities like Vector, Matrix, Plane, Sphere,
  etc. into the oglplus/math/ subdirectory.
- Work-in-progress, some parts won't compile.

+++ include/oglplus/vertex_attrib.hpp
@@ -388,7 +388,7 @@ public:
 	 *  @glfunref{GetAttribLocation}
 	 */
 	VertexArrayAttrib(VertexAttribSlot location)
-	 : ProgVarCommonOps<tag::VertexAttrib>(VertexAttribLoc(location))
+	 : ProgVarCommonOps<tag::VertexAttrib>(VertexAttribLoc(GLint(location)))
 	{ }
 
 	/// References the vertex attribute array at @p location
